Add identity operation
Generating it is never useful, but in optimizations it *is* useful to
identify that some things are equal to others, and then we need an
expressin to substitute the old one with. It could be done in a
different pass that runs in the full Gen monad, but this is a bit tidier
at the optimizer side, I think.