Implement elimination of string equality check
If the operands are the same, then they are equal. However, in my test
program, they are not the same -- they are two different string
constants that happen to be equal, so this optimization does not kick
in. I should implement common subexpression elimination (which is a bad
name in this case actually). Then one of the variables will turn into a
reference to the other, and then the equality check will annihilate.